Change wraplen and wrapmin to use usize instead of u8

+++ src/cli.rs
@@ -68,7 +68,7 @@ pub fn get_cli_args(matches: Option<ArgMatches>) -> OptionArgs {
         print: get_flag(&arg_matches, "print"),
         fail_on_change: get_flag(&arg_matches, "fail-on-change"),
         wrap,
-        wraplen: arg_matches.get_one::<u8>("wraplen").copied(),
+        wraplen: arg_matches.get_one::<usize>("wraplen").copied(),
         wrapmin: None,
         tabsize: arg_matches.get_one::<u8>("tabsize").copied(),
         tabchar,
